Fix freeze on entering full screen window mode
On macOS Mojave, window freeze when entering full screen window mode.
If the window is frozen, clicking outside the app allows it to complete
the transition to full screen mode. So even just clicking on another
icon in the Dock allows it to continue (Issue MiMo42/MMTabBarView#63).

Brendan Duddridge (@brendand) noticed that the problem has to do with
MMTabBarView’s -viewDidEndLiveResize. Removing the animation works
around it.