Match declaration & implementation of delegate
In commit 9e03810415250ed7e55cc8576a6373c1f1d22bf3, I changed the implementation of the delegate property to match the "weak" declaration. This was braking the demo app, so I now revert the implementation to "strong", and change the declaration so it matches the implementation.

Note the the client app should make sure that the delegate does not have
a strong reference to MMTabBarView, or there will be a retain cycle.

Author : @jeromectm - commits 9e03810415250ed7e55cc8576a6373c1f1d22bf3
and 3d911232d37ab221ecf2ce9a06b27fb960f7934b