Fix overflow button not showing with Mojave style
Fix issue: when launched with Mojave style enabled, the Demo app does
not show the symbol of the overflow menu when numerous tabs get opened.
When one switch to Sierra style and switch back to Mojave, things get
back to expected behavior.

Copy and paste a method from Sierra style.